+++ R/is.xifti.R
@@ -253,11 +253,12 @@ is.xifti_meta <- function(x) {
     }
     # [TO DO]: Cleanup.
     if (mean(x$cortex$medial_wall_mask$left) < .5) {
-      warning(paste(
-        "Warning: FALSE values in left medial wall mask should indicate",
-        "the medial wall vertices. But, there were more FALSE values than TRUE."
-      ))
-      #return(FALSE)
+      if (!(!is.null(x$cifti$small_ROI) && isTRUE(x$cifti$small_ROI))) {
+        warning(paste(
+          "Warning: FALSE values in left medial wall mask should indicate",
+          "the medial wall vertices. But, there were more FALSE values than TRUE."
+        ))
+      }
     }
   }
   if (!is.null(x$cortex$medial_wall_mask$right)) {
@@ -266,11 +267,12 @@ is.xifti_meta <- function(x) {
     }
     # [TO DO]: Cleanup.
     if (mean(x$cortex$medial_wall_mask$right) < .5) {
-      warning(paste(
-        "Warning: FALSE values in right medial wall mask should indicate",
-        "the medial wall vertices. But, there were more FALSE values than TRUE."
-      ))
-      #return(FALSE)
+      if (!(!is.null(x$cifti$small_ROI) && isTRUE(x$cifti$small_ROI))) {
+        warning(paste(
+          "Warning: FALSE values in right medial wall mask should indicate",
+          "the medial wall vertices. But, there were more FALSE values than TRUE."
+        ))
+      }
     }
   }
   if ((!is.null(x$cortex$medial_wall_mask$left)) && !is.null(x$cortex$medial_wall_mask$right)) {
@@ -519,15 +521,21 @@ is.xifti <- function(x, messages=TRUE) {
         return(FALSE)
       }
 
+      vl_show_max <- 20
+
       if (!is.null(x$data$cortex_left)) {
         for (ii in seq_len(ncol(x$data$cortex_left))) {
           all_labels <- unique(data_mat[,ii])
           valid_label <- all_labels %in% x$meta$cifti$labels[[ii]]$Key
           if (!all(valid_label)) {
+            vl_show <- which(!valid_label)
+            vl_trnc <- length(vl_show) > vl_show_max
+            if (vl_trnc) { vl_show <- vl_show[seq(vl_show_max)] }
             message(paste(
               "These label values in the left cortex data column", ii, 
               "are not in the corresponding label table:\n\t",
-              paste(all_labels[!valid_label], collapse=", "), "\n"
+              paste(all_labels[vl_show], collapse=", "), 
+              ifelse(vl_trnc, "[TRUNCATED]", ""), "\n"
             ))
             return(FALSE)
           }
@@ -539,10 +547,14 @@ is.xifti <- function(x, messages=TRUE) {
           all_labels <- unique(data_mat[,ii])
           valid_label <- all_labels %in% x$meta$cifti$labels[[ii]]$Key
           if (!all(valid_label)) {
+            vl_show <- which(!valid_label)
+            vl_trnc <- length(vl_show) > vl_show_max
+            if (vl_trnc) { vl_show <- vl_show[seq(vl_show_max)] }
             message(paste(
               "These label values in the right cortex data column", ii, 
               "are not in the corresponding label table:\n\t",
-              paste(all_labels[!valid_label], collapse=", "), "\n"
+              paste(all_labels[vl_show], collapse=", "), 
+              ifelse(vl_trnc, "[TRUNCATED]", ""), "\n"
             ))
             return(FALSE)
           }
@@ -554,10 +566,14 @@ is.xifti <- function(x, messages=TRUE) {
           all_labels <- unique(data_mat[,ii])
           valid_label <- all_labels %in% x$meta$cifti$labels[[ii]]$Key
           if (!all(valid_label)) {
+            vl_show <- which(!valid_label)
+            vl_trnc <- length(vl_show) > vl_show_max
+            if (vl_trnc) { vl_show <- vl_show[seq(vl_show_max)] }
             message(paste(
               "These label values in the subcortex data column", ii, 
               "are not in the corresponding label table:\n\t",
-              paste(all_labels[!valid_label], collapse=", "), "\n"
+              paste(all_labels[vl_show], collapse=", "), 
+              ifelse(vl_trnc, "[TRUNCATED]", ""), "\n"
             ))
             return(FALSE)
           }