+++ R/make_xifti_components.R
@@ -532,6 +532,6 @@ make_surf <- function(surf, expected_hemisphere=NULL) {
 
 #' @rdname make_surf
 #' @export
-gifti_to_surf <- function(surf, expected_hemisphere=NULL){
+read_surf <- function(surf, expected_hemisphere=NULL){
   make_surf(surf=surf, expected_hemisphere=expected_hemisphere)
 }