+++ R/smooth_cifti.R
@@ -15,7 +15,7 @@
 #' 
 #' @inheritSection Connectome_Workbench_Description Connectome Workbench Requirement
 #' 
-#' @param x The CIFTI file name or \code{"xifti"} object to resample.
+#' @param x The CIFTI file name or \code{"xifti"} object to smooth.
 #' @param cifti_target_fname File name for the smoothed CIFTI. If
 #'  \code{NULL}, will be written to "smoothed.d*.nii" in the current working
 #'  directory if \code{x} was a CIFTI file, and in a temporary directory if